quick_exit tests: Do not use C++ headers
If C++ headers such as <cstdlib> or <thread> are used, GCC 6
will include /usr/include/stdlib.h (instead of stdlib/stdlib.h
in the glibc source directory), and this turns up as a make
dependency.  An implicit rule will kick in and make will try to
install stdlib/stdlib.h as /usr/include/stdlib.h because the
target is out of date.

This commit switches to <stdlib.h> and <pthread.h> instead of
<cstdlib> and <thread>.