Fix math.h comment about bits/mathdef.h.
math.h has a comment about definitions from <bits/mathdef.h>.  This
comment is in the wrong place in math.h, far below the inclusion of
<bits/mathdef.h>.  It was originally above the inclusion, but the
inclusion was moved by

1998-11-05  Ulrich Drepper  <drepper@cygnus.com>

        * math/math.h: Unconditionally include bits/mathdef.h.  Declare
        long double functions only if __NO_LONG_DOUBLE_MATH is not
        defined.
        [...]

without moving the comment.  Furthermore, the comment refers
incorrectly to FLT_EVAL_METHOD and DECIMAL_DIG, which are actually
<float.h> macros, and INFINITY, which is in <bits/inf.h>.

This patch moves the comment back above the include it refers to and
removes the description of macros not defined by the header.

Tested for x86_64 and x86 (testsuite, and that installed stripped
shared libraries are unchanged by the patch).

	* math/math.h: Move comment about <bits/mathdef.h> definitions
	above inclusion of <bits/mathdef.h>.  Do not mention
	FLT_EVAL_METHOD, INFINITY or DECIMAL_DIG in that comment.